requirements.txt
# Core
Django==4.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
django-filter==23.5
python-dotenv==1.0.1

# DB
psycopg2-binary==2.9.9
psycopg2==2.9.10

# Image Manipulation
Pillow==11.2.1

# Setup
setuptools==80.8.0

# Envs
django-environ==0.12.0
